moved stepconfirm to parent folder to be shared by similar workflows
